Initial reworking of network code to utilise new TLS API in node 0.4+.
I'm not entirely sure I have the error checking right/sufficient.

The method of queueing pending messages whilst connecting has been changed to an array which the data is pushed/shifted from as it is no longer possible to add a listener to trigger the write because the socket will change as it is recreated.