Update README.md
Updated the wiki by adding the information how to convert a string into a hexadecimal string. I was doing it wrong before and it took me quite some time to find my stupid error, so I hope this information might help others not to make the same mistake.

By the way, thanks a lot for publishing this very useful module!