Don't blindly reject the deferredConnection when an error occurs
If a new connection is established before the error from the previous connection is thrown then a deferredConnection will exist and it will happily reject it, upsetting the new connection flow. In fact we should check that the error occurred on the in progress socket (`this.socket`) before we destroy that.

Otherwise just handle it as a regular socket error.