Fixed connection for 0.11 and older versions
Fixes #122.

node-0.11 changes how TLS connections handle sockets when connecting.

Discovered in #65, it's necessary to attach the error handler before the socket establishes a connection. In some cases the socket errors before the TLS#connect method returns, so the error handler hasn't been attached yet and an exception is raised.