The tests didn't really reflect real-world cases
It's not likely that both PFX and PEM data will be passed in simultaneously and it is usually an error to do so. Something that may be raised in future behaviour and I don't want to have to change these tests.

Also PFX parsing is really slow so only do it when absolutely necessary.