Prevent wakeup being emitted on connect.
We know that APNS will send a stream limit immediately after connection
and given the current state of node-http2 we have to start with an
initial stream limit of 0 until that comes through. As such it is
redundant to emit wakeup immediately after connection because we won't
be able to do anything.

In addition, if we do go back to having an infinite stream limit
initially we don't want to flood a single endpoint with streams when we
could distribute it across multiple. (Imagine a call to
`pushNotification` with 10,000 tokens all piling onto a single
connection).