Manage the stream counts within Endpoint
Using the built-in available stream count was not suitable as it is
managed on a different schedule to that which we require. `createStream`
would not increment he used stream count until a frame was actually sent
whereas we want to identify a stream as in-use as soon as it is acquired
so we shall maange a local count.