Ensure the stream is allocated an ID
We need to ensure a stream has entered the `_streamIds` array so we can
loop over it if an error occurs.

If an error occurs we need to clean up old streams by bubbling the error
through to them, allowing their corresponding promise to be rejected. If
the error doesn't make it through the promise will stall forever.

An error occurring between a stream being created (popping it from the
client queue) and having data written (when it would normally be
allocated an ID) then it can end up in limbo because looping through
`streamIds` would not hit that stream.