Ensure we reset the failure count
If we don't reset the failure count then the pending notification queue
will stall without warning. Instead it seems fair that if the
application continues to send notifications then we give it another
shot and attempt 3 more times.

This has the effect of never really "hard" failing but all notifications
are handled and the application can take responsibility for abandoning
the worker if they are monitoring the errors correctly.